The prior "fixed an obviously wrong assignment" change was bad
obj.opId was set to the correct value a couple of lines earlier.  Don't
know why the "identity assignment" was there but this change removes it
to avoid future confusion.

This fixes both of the console regressions: hover on stream not
reporting tuple cnt and hover on Union not reporting correct summed
output tuple count.