simplify: remove supplemental-model.xml stuff
leaving this out until we're confident we need to
"correct" inaccurate pom info from external dependencies
(used for their content in in generated DEPENDENCIES files).

can always revert this commit.

P.S. not sure the poms for the external deps in the connectors
was reviewed - i.e., whether supplemental-model.xml was complete.