add release validation check_jars.sh
Initial version of tools to help check that a release contains the
appropriate jars and they have the required metadata content.

This version lacks integration with Nexus - the real target of this
validation checking.  If there's a tool that can download all of the
jars present in a nexus staging area then this tool can be run against
that downloaded tree.